make mappers

+++ app/src/main/java/com/example/weatherapp/data/remote/WeatherDataDto.kt
@@ -11,7 +11,7 @@ data class WeatherDataDto(
     @field:Json(name = "pressure_msl")
     val pressures: List<Double>,
     @field:Json(name = "windspeed_10m")
-    val windSpeeds: List<Int>,
+    val windSpeeds: List<Double>,
     @field:Json(name = "relativehumidity_2m")
     val humidities: List<Double>,
 )
